Unit Test harness temporary workaround fix
Temporarily workaround fixed an issue where an incorrect value would cause the test to bug out while trying to display the debug values. The int values attempting to display were to large to be converted to a string, fix is to not attempt to show debug values in some tests.